requirements.txt
asgiref==3.8.1
colorama==0.4.6
coverage==7.4.4
Django==4.1.13
django-cors-headers==4.3.1
djangorestframework==3.15.1
djongo==1.3.6
dnspython==2.6.1
exceptiongroup==1.2.1
iniconfig==2.0.0
packaging==24.0
pluggy==1.5.0
pymongo==4.6.3
pytest==8.1.1
pytest-cov==5.0.0
pytest-django==4.8.0
sqlparse==0.2.4
tomli==2.0.1
typing-extensions==4.11.0
tzdata==2024.1
